refactor: Extract setup message initialization logic into separate module
- Create new initSetupMessage.ts to handle DKLS keygen setup message logic
- Move setup message creation and download/upload functionality from startDklsKeygen
- Simplify startDklsKeygen by delegating setup message initialization
- Improve code modularity and separation of concerns
